Lift masked-tile Cond hoist into 021_hoist_staged_loads_above_mask
Splits the 1156-line `020_stage_inputs.py` along its natural fence:
staging policy (admit Loads, build Sources, wrap StageBundle) stays
in 020; the rearrangement that hoists the cooperative-load K-pipeline
above the masked-tile boundary Cond moves to a new focused pass.

`020_stage_inputs._process_scope`'s Cond branch (88 lines) shrinks to
a uniform transparent descent (21 lines); five helpers move verbatim
(`_is_k_pipeline_stmt`, `_contains_stage_bundle`, `_collect_smem_names`,
`_guard_unsafe_loads`, `_guard_unsafe_loads_in_body`). The new pass
runs `Body.map(_lift_if_match)`, finds `Cond(decoded < bound, body=...)`
shapes whose body contains a `StageBundle`, partitions the body into
K-pipeline stmts (hoisted) and the rest (kept inside the Cond), and
wraps un-staged gmem Loads referencing gated vars in inner Conds.

Side-effect fix: the previous early-exit in 020's Cond branch
discarded the recursive descent's rewrites when the hoist precondition
failed. The split preserves the inner staging unconditionally.

Verified end-to-end:
- targeted: test_masked_tile (8), test_dynamic_shapes (12),
  test_knob_pinning fm26/fn26/article_exact_fm26 (8),
  test_qwen_lmhead_variant_compiles_within_budget (1),
  stage_inputs unit tests (8)
- full compiler suite serial: 1078 passed, 2 skipped, 0 failures
- make lint clean
