+++ deplodock/compiler/pipeline/ARCHITECTURE.md
@@ -439,7 +439,7 @@ ignores the prefix itself — it's only for ordering readability.
 | `frontend/optimization/`   | `compose_indexmaps`: collapse chains of single-source / single-consumer `IndexMapOp` into one coord_map — prevents trivial layout kernels from blocking fusion. |
 | `loop/lifting/`            | `lift_*` rules wrap each surviving tensor primitive (elementwise / reduce / indexmap / gather) in a trivial one-op `LoopOp`. |
 | `loop/fusion/`             | `split_shared_indexmap` (runs first) fuses a pure-indexmap `LoopOp` that fans out to ≥2 consumers into **all** of them in one rewrite — it inlines the producer's body into each consumer (reusing `splice_loop_ops`) and dissolves the producer via a single multi-output `Graph.splice` (`output={consumer_id: fused_id}`); a consumer the splicer can't take falls back to a private copy. Then `merge_loop_ops` splices the remaining adjacent single-consumer `LoopOp` pairs via `ir/loop/splicer.py::splice_graph`. The split is what lets the scalar-constant broadcasts torch.export folds into mask/RoPE scaffolding fold into their consumers (inlined as `float x = 0.0f;` literals) instead of surviving as standalone copy kernels (full Qwen3-Embedding-0.6B: 394 → 337 CUDA kernels). `dedup_loads` then drops identical `(input, index)` Loads within each fused body. Finally `stamp_loop_names` stamps `LoopOp.name` via `provenance.name_for` (e.g. `k_rms_norm_3f2a1b`) — runs last so the structural hash reflects the final fused body; the Tile dialect just forwards the name onto each emitted `TileOp` (and every dialect below copies it through unchanged). Shared `is_pure_indexmap` / `rename_write_output` helpers live in `_helpers.py`. |
-| `lowering/tile/`           | Tile-IR structural passes — Stage formation, transport (cp.async / TMA), double-buffering, pipelining, smem padding. Order: `partition_loops` → `gate_splitk_residual` → `stage_inputs` → `swizzle_blocks` (default-on
+| `lowering/tile/`           | Tile-IR structural passes — Stage formation, transport (cp.async / TMA), double-buffering, pipelining, smem padding. Order: `partition_loops` → `gate_splitk_residual` → `stage_inputs` → `hoist_staged_loads_above_mask` (lifts a masked-tile boundary `Cond(decoded < bound, ...)` from `010_partition_loops`: any K-pipeline stmt — `StageBundle` itself, plus `SerialTile` / `StridedTile` whose subtree carries one — is hoisted ABOVE the Cond so the cooperative load fires on every CTA thread (TMA's elected issuer / cp.async's full-CTA fan-out would otherwise be gated out). Un-staged gmem Loads in the hoisted body whose index references a gated var are wrapped in an inner `Cond(predicate, body=cone)` covering their forward SSA cone. Skips `==` Conds (the SPLITK invariant-compute guard) and bare Conds with no staged transport. Deterministic, no knob — split out of `020_stage_inputs` so the staging walk is uniform and the Cond-shape rewrite is focused) → `swizzle_blocks` (default-on
 L2-friendly CTA swizzle for matmul-priority TileOps — stamps `GridTile.swizzle_group_m = DEPLODOCK_GROUP_M`,
 default 8, so the renderer emits a Triton-canonical `blockIdx.x` remap; identifies matmul kernels via
 `TileOp.knobs` rather than the axis-suffix convention because the body-normalizer renames axes by the time
